Mavenize project, and improve code readibility
