--- TimeIsMoney2/bin/plugin.yml
@@ -1,17 +1,0 @@
-name: TimeIsMoney
-depends: [Vault]
-soft-depends: [Essentials]
-author: Linus122
-main: de.Linus122.TimeIsMoney.Main
-version: 1958
-description: Gives money for online time
-commands:
-   timeismoney:
-      description: Reloads the Config
-      aliases: tim
-      usage: /<command>
-      permission: tim.reload
-   atm:
-      description: Opens the atm for other player
-      usage: /<command>
-      permission: tim.admin
