--- TimeIsMoney2/src/de/Linus122/TimeIsMoney/ActionBarUtils.java
@@ -1,7 +1,0 @@
-package de.Linus122.TimeIsMoney;
-
-import org.bukkit.entity.Player;
-
-public interface ActionBarUtils {
-	void sendActionBarMessage(Player p, String message);
-}
