+++ src/main/java/de/Linus122/TimeIsMoney/Payout.java
@@ -1,0 +1,12 @@
+package de.Linus122.TimeIsMoney;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Payout {
+    double payout_amount = 0;
+    double max_payout_per_day = 0;
+    String permission = "";
+    int chance = 0;
+    List<String> commands = new ArrayList<>();
+}
