+++ src/main/java/v1_12_R1/NBTUtils.java
@@ -1,0 +1,18 @@
+package v1_12_R1;
+
+import de.Linus122.TimeIsMoney.ActionBarUtils;
+import net.minecraft.server.v1_12_R1.ChatMessageType;
+import net.minecraft.server.v1_12_R1.IChatBaseComponent;
+import net.minecraft.server.v1_12_R1.PacketPlayOutChat;
+import org.bukkit.craftbukkit.v1_12_R1.entity.CraftPlayer;
+import org.bukkit.entity.Player;
+
+import static de.Linus122.TimeIsMoney.Utils.CC;
+
+public class NBTUtils implements ActionBarUtils {
+    public void sendActionBarMessage(Player p, String message) {
+        IChatBaseComponent icbc = IChatBaseComponent.ChatSerializer.a("{\"text\": \"" + CC(message) + "\"}");
+        PacketPlayOutChat bar = new PacketPlayOutChat(icbc, ChatMessageType.GAME_INFO);
+        ((CraftPlayer) p).getHandle().playerConnection.sendPacket(bar);
+    }
+}
